requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-braces==1.15.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.4
wrapt==1.15.0
